+++ apps/webservice/src/app/api/github/webhook/workflow/handler.ts
@@ -1,23 +1,9 @@
 import type { WorkflowRunEvent } from "@octokit/webhooks-types";
 
-import {
-  and,
-  count,
-  desc,
-  eq,
-  takeFirst,
-  takeFirstOrNull,
-} from "@ctrlplane/db";
+import { and, eq, takeFirstOrNull } from "@ctrlplane/db";
 import { db } from "@ctrlplane/db/client";
 import * as schema from "@ctrlplane/db/schema";
-import {
-  cancelOldReleaseJobTriggersOnJobDispatch,
-  createReleaseJobTriggers,
-  dispatchReleaseJobTriggers,
-  isPassingAllPolicies,
-  onJobCompletion,
-} from "@ctrlplane/job-dispatch";
-import { logger } from "@ctrlplane/logger";
+import { onJobCompletion, onJobFailure } from "@ctrlplane/job-dispatch";
 import { JobStatus } from "@ctrlplane/validators/jobs";
 
 type Conclusion = Exclude<WorkflowRunEvent["workflow_run"]["conclusion"], null>;
@@ -59,86 +45,6 @@ const getJob = async (externalId: number, name: string) => {
     .from(schema.job)
     .where(eq(schema.job.id, uuid))
     .then(takeFirstOrNull);
-};
-
-const dispatchJobsForNewerRelease = async (releaseId: string) => {
-  const releaseJobTriggers = await db
-    .select()
-    .from(schema.releaseJobTrigger)
-    .innerJoin(schema.job, eq(schema.releaseJobTrigger.jobId, schema.job.id))
-    .where(
-      and(
-        eq(schema.releaseJobTrigger.releaseId, releaseId),
-        eq(schema.job.status, JobStatus.Pending),
-      ),
-    )
-    .then((rows) => rows.map((r) => r.release_job_trigger));
-
-  return dispatchReleaseJobTriggers(db)
-    .releaseTriggers(releaseJobTriggers)
-    .filter(isPassingAllPolicies)
-    .then(cancelOldReleaseJobTriggersOnJobDispatch)
-    .dispatch();
-};
-
-const maybeRetryJob = async (job: schema.Job) => {
-  const jobInfo = await db
-    .select()
-    .from(schema.releaseJobTrigger)
-    .innerJoin(
-      schema.release,
-      eq(schema.releaseJobTrigger.releaseId, schema.release.id),
-    )
-    .innerJoin(
-      schema.deployment,
-      eq(schema.release.deploymentId, schema.deployment.id),
-    )
-    .where(eq(schema.releaseJobTrigger.jobId, job.id))
-    .then(takeFirstOrNull);
-
-  if (jobInfo == null) return;
-
-  const latestRelease = await db
-    .select()
-    .from(schema.release)
-    .where(eq(schema.release.deploymentId, jobInfo.deployment.id))
-    .orderBy(desc(schema.release.createdAt))
-    .limit(1)
-    .then(takeFirst);
-
-  if (latestRelease.id !== jobInfo.release.id)
-    return dispatchJobsForNewerRelease(latestRelease.id);
-
-  const releaseJobTriggers = await db
-    .select({ count: count() })
-    .from(schema.releaseJobTrigger)
-    .where(eq(schema.releaseJobTrigger.releaseId, jobInfo.release.id))
-    .then(takeFirst);
-
-  const { count: releaseJobTriggerCount } = releaseJobTriggers;
-
-  if (releaseJobTriggerCount >= jobInfo.deployment.retryCount) return;
-
-  const createTrigger = createReleaseJobTriggers(db, "retry")
-    .releases([jobInfo.release.id])
-    .environments([jobInfo.release_job_trigger.environmentId]);
-
-  const trigger =
-    jobInfo.release_job_trigger.causedById != null
-      ? await createTrigger
-          .causedById(jobInfo.release_job_trigger.causedById)
-          .insert()
-      : await createTrigger.insert();
-
-  await dispatchReleaseJobTriggers(db)
-    .releaseTriggers(trigger)
-    .then(cancelOldReleaseJobTriggersOnJobDispatch)
-    .dispatch()
-    .then(() =>
-      logger.info(
-        `Retry job for release ${jobInfo.release.id} and resource ${jobInfo.release_job_trigger.resourceId} created and dispatched.`,
-      ),
-    );
 };
 
 export const handleWorkflowWebhookEvent = async (event: WorkflowRunEvent) => {
@@ -188,6 +94,6 @@ export const handleWorkflowWebhookEvent = async (event: WorkflowRunEvent) => {
       set: { value: links },
     });
 
-  if (status === JobStatus.Failure) await maybeRetryJob(job);
+  if (status === JobStatus.Failure) await onJobFailure(job);
   if (status === JobStatus.Completed) return onJobCompletion(job);
 };
